Update carbon reduction & Add Naver Clova OCR

+++ backend/src/main/java/com/ssafy/wtd/backend/dto/charge/ChargeConfirmRes.java
@@ -2,7 +2,6 @@ package com.ssafy.wtd.backend.dto.charge;
 
 import lombok.Getter;
 import lombok.Setter;
-import java.time.LocalDateTime;
 
 @Getter
 @Setter
@@ -18,6 +17,6 @@ public class ChargeConfirmRes { // 최종 저장 후 결과
         private String stationId;
         private float chargedKwh;
         private int chargingCost;
-        private int durationMin;
+        private float durationMin;
     }
 }
